+++ npm/packages/commons-firebase-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0](https://github.com/aimarchirico/commons/compare/commons-firebase-client-v1.0.2...commons-firebase-client-v1.1.0) (2026-07-17)
+
+
+### Features
+
+* **tools:** add root:fix task and fix docs pathing ([8d75182](https://github.com/aimarchirico/commons/commit/8d75182043713d6d389532ed60c783781df2cdad))
+
 ## [1.0.2](https://github.com/aimarchirico/commons/compare/commons-firebase-client-v1.0.1...commons-firebase-client-v1.0.2) (2026-07-06)